requirements.txt
# Web 框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
aiohttp==3.9.5

# 数据库
pymongo==4.6.0
motor==3.3.2
beanie==1.23.6
redis==5.0.1
pymilvus==2.3.3

# 消息队列
kafka-python==2.0.2

# 大模型和 RAG
langchain==0.3.0
langchain-community==0.3.0
langchain-openai==0.2.14
ollama==0.1.6

# Embedding 模型
sentence-transformers==2.2.2
transformers==4.41.2
torch==2.2.2
safetensors>=0.4.3

# Re-ranker 模型
FlagEmbedding==1.2.10

# 文档处理
pypdf2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6

# 工具类
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
numpy==1.24.3

# 日志
loguru==0.7.2

